requirements.txt
jinja2==3.1.4
python-hcl2==4.3.5
pyyaml==6.0.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
openai>=1.12.0
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
pydantic-settings==2.3.4
passlib[bcrypt]==1.7.4
httpx==0.28.1

bcrypt<5
